Corrected location of static inlines, modified Makefile.depend to pull in common headers.
