+++ spinnaker_components/common/dimensional-io.h
@@ -43,9 +43,29 @@ typedef struct filtered_input_buffer {
 } filtered_input_buffer_t;
 
 //! Create and initialise an input buffer, and zero the accumulator
-static inline filtered_input_buffer_t* input_buffer_initialise( uint d_in );
+static inline filtered_input_buffer_t* input_buffer_initialise( uint d_in ) {
+  // Create the buffer on the heap
+  filtered_input_buffer_t *buffer = spin1_malloc(
+    sizeof( filtered_input_buffer_t )
+  );
+  buffer->d_in = d_in;
+
+  // Initialise the buffer accumulator and values
+  buffer->accumulator = spin1_malloc( sizeof( accum ) * d_in );
+  buffer->filtered = spin1_malloc( sizeof( accum ) * d_in );
+}
 
 //! Filter an input buffer and zero the accumulator
-static inline void input_buffer_step( filtered_input_buffer_t *buffer );
+static inline void input_buffer_step( filtered_input_buffer_t *buffer ) {
+  // For each input dimension perform filtering and zero the accumulator
+  for( uint d = 0; d < buffer->d_in; d++ ){
+    // Perform the filtering
+    buffer->filtered[d] *= buffer->filter;
+    buffer->filtered[d] += buffer->accumulator[d] * buffer->n_filter;
+
+    // Zero the accumulator
+    buffer->accumulator[d] = 0;
+  }
+}
 
 #endif
